freerds: cleanup RPC stubs

+++ freerds/core/app_context.h
@@ -27,17 +27,17 @@
 
 struct application_context 
 {
-	wListDictionary *connections;
-	long connectionId;
+	wListDictionary* connections;
+	UINT32 connectionId;
 };
 
 typedef struct application_context rdsAppContext;
 
 void app_context_init();
 void app_context_uninit();
-long app_context_get_connectionid();
-void app_context_add_connection(rdsConnection *connection);
-void app_context_remove_connection(long id);
-rdsConnection* app_context_get_connection(long id);
+UINT32 app_context_get_connectionid();
+void app_context_add_connection(rdsConnection* connection);
+void app_context_remove_connection(UINT32 id);
+rdsConnection* app_context_get_connection(UINT32 id);
 
 #endif /* FREERDS_APPCONTEXT_H */
